requirements.txt
affine==2.4.0
attrs==25.1.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2025.1.31
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
geopandas==1.0.1
h5py==3.12.1
imageio==2.37.0
lazy_loader==0.4
networkx==3.4.2
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
rasterio==1.4.3
scikit-image==0.25.1
scipy==1.15.1
shapely==2.0.7
six==1.17.0
soupsieve==2.6
tifffile==2025.1.10
tzdata==2025.1
